requirements.txt
Flask
pandas
scikit-surprise